creating section Q&A
adding section,
adding icons,
adding accordion,
adding ShareButton
