+++ messages/uk.json
@@ -115,6 +115,7 @@
         "answer5": "На нашому сайті ви знайдете інформацію про вже реалізовані проекти та команди, які стояли за їхнім створенням, а також детальні профілі кожного учасника команди з описом їхніх навичок, технологічного стеку, посиланнями на соціальні мережі та контактними даними. ",
         "buttonShare": "Поділитися",
         "button": "button"
+    },
     "Buttons": {
         "order": "Замовити",
         "joinUs": "Долучитися",
