+++ src/components/qAsection/QaSection.tsx
@@ -11,7 +11,7 @@ export const QaSection = () => {
         <>
             <Section className="relative">
                 <TitleQA />
-                <ButtonShare className="absolute float-right right-[60px] top-[23px] mb-[64px] my-[8px]">
+                <ButtonShare className="absolute right-[60px] top-[23px] mb-[64px] my-[8px]">
                     {getTranslation("buttonShare")}
                 </ButtonShare>
                 <QuestionsList />
